🛡️ Sentinel: Finalize security hardening and fix Maestro CI compatibility
- Consolidate VpnStatus enum into a single shared location with displayText.
- Allow loopback cleartext traffic in debug builds for Maestro E2E driver.
- Enforce strict production hardening (usesCleartextTraffic=false, allowBackup=false).
- Fix manifest merger by providing actual values for replaced attributes in debug manifest.
- Fix git submodule metadata error by adding .gitmodules.
- Migrate GeoIP lookups to HTTPS via ipwho.is.
- Upgrade AGP to 8.2.1 for Java 21 stability.
