+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
         android:usesCleartextTraffic="false"
         android:banner="@drawable/tv_banner"
         tools:targetApi="31"
-        tools:replace="android:allowBackup,android:networkSecurityConfig,android:usesCleartextTraffic,android:theme,android:name,android:label">
+        tools:replace="android:allowBackup,android:networkSecurityConfig,android:usesCleartextTraffic">
         
         <!-- Mobile Activity -->
         <activity
